fix(signal_engine): bound DFS stacks (OOB), widen slot_count guard, compile-depth limit, prev-batch, classify refresh, oversize diagnostics
FW1: bounds-guard DFS reachability stacks in NodePool::rebuild_execution_order + gc_unreachable_nodes (+child<node_count guard) — was OOB write on dense DAGs.
FW2: widen OutputDeps::slot_count uint8_t->uint16_t — 'slot_count < MAX_LIVE_SLOTS(256)' was always-true on WASM, guard never fired.
FW3: MAX_COMPILE_DEPTH=64 RAII guard in compile_expr — was unbounded recursion on RP2040 stack inside tick().
FW4: route prev-feedback (no defstate) outputs through sequential batch path — fast path read stale prev() per sample (prev.md §1.4/1.6).
FW5: classify_outputs() after recompile in on_cell_changed — output_class/input_mask went stale.
FW6: tokenizer + serial extract_string emit 'program too large' diagnostics instead of silent truncation.
CF7: useq_active_diagnostics() stub returns '[]' not '{}' (diagnostics.md §4.2 array contract).

Verified: build:wasm clean, 12/12 wasmInterpreter + 8/8 new wasmResourceSafety tests pass.
NOTE: firmware-only serial_protocol.cpp change not PlatformIO-compiled here (backward-compatible defaulted param) — needs a hardware build to fully verify.

+++ uSEQ/src/signal_engine/node_pool.h
@@ -89,9 +89,14 @@ struct OutputDeps {
     SymbolID cells[MAX_OUTPUT_DEPS] = {};
     uint8_t count = 0;
 
-    // Live-edit slot indices referenced by this output's graph
+    // Live-edit slot indices referenced by this output's graph.
+    // slot_count must be wide enough to actually reach MAX_LIVE_SLOTS: on the
+    // WASM build MAX_LIVE_SLOTS == 256, which a uint8_t can never represent, so
+    // a `slot_count < MAX_LIVE_SLOTS` guard with a uint8_t counter is always
+    // true and never fires. uint16_t covers both the firmware (16) and WASM
+    // (256) capacities.
     uint16_t slots[MAX_LIVE_SLOTS] = {};
-    uint8_t slot_count = 0;
+    uint16_t slot_count = 0;
 
     void clear();
     void add(SymbolID sym);
